Adds a namespace to all objects (#1)
TL;DR
-----

Provides an overlay that adds a namespace to all objects

Details
-------

Implements an overlay that adds a namespace to all objects in
the processed manifests that don't already have one. It's a fairly
blunt tool since it specifies one namespace that is used for all
objects that aren't in one.